refactor(WebAPI): changed namespace to AppHost
