+++ tests/BaseTests/Common/ReaparrWebApplicationFactory.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.AspNetCore.TestHost;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Reaparr.WebAPI;
+using Reaparr.AppHost;
 
 namespace Reaparr.BaseTests;
